Ensure moving to next neigbour does not add duplicate staff member
Ensure that when swapping a staff member on a shift for another staff
member from the available set, we do not add one that is already on the
shift.

Also, when creating copy of shift to modify, copy the staff collection
too (don't just assign it) then ensure the copy of the shift overwrites
the original one in the array.